# Internship/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Registration System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event operations
  - name: registrations
    description: Registration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: upcoming
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to only upcoming events (event date in the future)
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [date, name]
          description: Sort events by date or name
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Event name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{eventId}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [registrations]
      summary: List active registrations for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register a user for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already registered or event is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /registrations/{registrationId}:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/stats/summary:
    get:
      operationId: getEventsSummary
      tags: [events]
      summary: Get summary stats for all events
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Event:
      type: object
      required: [id, name, totalSeats, availableSeats, eventDate, totalRegistrations, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        totalSeats:
          type: integer
        availableSeats:
          type: integer
        totalRegistrations:
          type: integer
        eventDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    EventInput:
      type: object
      required: [name, totalSeats, eventDate]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        totalSeats:
          type: integer
          minimum: 1
        eventDate:
          type: string
          format: date-time

    Registration:
      type: object
      required: [id, eventId, userName, status, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userName:
          type: string
        status:
          type: string
          enum: [active, cancelled]
        registeredAt:
          type: string
          format: date-time
        cancelledAt:
          type: ["string", "null"]

    RegistrationInput:
      type: object
      required: [userName]
      properties:
        userName:
          type: string
          minLength: 1

    EventsSummary:
      type: object
      required: [totalEvents, upcomingEvents, totalRegistrations, totalAvailableSeats]
      properties:
        totalEvents:
          type: integer
        upcomingEvents:
          type: integer
        totalRegistrations:
          type: integer
        totalAvailableSeats:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string